requirements.txt
behave
pytest
allure-pytest
allure-behave
httpx
jmespath
selenium
chromedriver-py